requirements.txt
asgiref==3.2.3
astroid==2.3.3
colorama==0.4.3
Django==2.0.2
django-filter==1.1.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.2.1
graphql-relay==0.4.5
iso8601==0.1.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
promise==2.2.1
psycopg2==2.8.4
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
